+++ CarcassonneScore.py
@@ -22,7 +22,7 @@ import argparse
 import sys
 import cgame
 
-__version__ = "0.3.4"
+__version__ = "0.4.2"
 
 def getargs():
     """