+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Add turn length vs turn number plot to `analysis/SampleAnalysis-SingleGame.ipynb`.
 * Reorder game initilization so that most of the info can be entered before the player order is determined.
+* Announce gameID and location at the start.
 
 ### 0.4.1 (18 November 2018)