+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Announce gameID and location at the start.
 * Status printout now shows time elapsed since end of previous turn.
 * In the status message, showing the number of tiles played/remaining and the time elapsed since the end of the previous turn is now controlled via options in the configuration file.
+* Add `requirements.txt` file.
 
 #### Bug fixes