+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 #### Bug fixes
 
 * Fixed tile counting bug when an Abbey is played.
+* Fixed missing Abbey score type when Expansion 5 is used.
+  This requres running the following command in sqlite3: `update expansions set scoretypes="Abbey" where expansionID=5;`
 
 ### 0.4.1 (18 November 2018)